+++ pid.h
@@ -1,14 +1,14 @@
 typedef struct 
 {
-    double windup_guard;
-    double proportional_gain;
-    double integral_gain;
-    double derivative_gain;
-    double prev_error;
-    double int_error;
-    double control;
+    float windup_guard;
+    float proportional_gain;
+    float integral_gain;
+    float derivative_gain;
+    float prev_error;
+    float int_error;
+    float control;
 } pid_type;
  
 
 void Pid_Reset(pid_type* pid);
-void Pid_Update(pid_type* pid, double current_error, double dt);
+void Pid_Update(pid_type* pid, float current_error, float dt);
